Rename jsx files to use jsx extension.
Organize imports and format on all jsx files.
