+++ client/src/Bit.jsx
@@ -8,16 +8,16 @@ import BitTagPills from './BitTagPills';
 export default function Bit(props) {
   const { bit = new Map(), upvote, downvote } = props;
   const header = (
-      <h3>
-        <Button bsStyle={getUpvoteButtonStyle(bit)} className="thumbs-up-button" onClick={() => upvote(bit.get('postId'))}>
-          <span className="glyphicon glyphicon-thumbs-up" />
-        </Button>
-        {bit.get('upvotes', 0) - bit.get('downvotes', 0) + bit.get('userVote', 0)}
-        <Button bsStyle={getDownvoteButtonStyle(bit)} className="thumbs-down-button" onClick={() => downvote(bit.get('postId'))}>
-          <span className="glyphicon glyphicon-thumbs-down" />
-        </Button>
-        {bit.get('title')}
-      </h3>
+    <h3>
+      <Button bsStyle={getUpvoteButtonStyle(bit)} className="thumbs-up-button" onClick={() => upvote(bit.get('postId'))}>
+        <span className="glyphicon glyphicon-thumbs-up" />
+      </Button>
+      {bit.get('upvotes', 0) - bit.get('downvotes', 0) + bit.get('userVote', 0)}
+      <Button bsStyle={getDownvoteButtonStyle(bit)} className="thumbs-down-button" onClick={() => downvote(bit.get('postId'))}>
+        <span className="glyphicon glyphicon-thumbs-down" />
+      </Button>
+      {bit.get('title')}
+    </h3>
   );
   return (
     <Panel header={header}>
@@ -30,7 +30,7 @@ export default function Bit(props) {
         </p>
         {bit.get('content')}
         <p>
-          <Link to={ '/bits/' + bit.get('postId') }>permalink</Link>
+          <Link to={'/bits/' + bit.get('postId')}>permalink</Link>
         </p>
       </div>
     </Panel>
@@ -38,7 +38,7 @@ export default function Bit(props) {
 }
 
 const getDownvoteButtonStyle = bit =>
-    bit.get('userVote') === USER_DOWNVOTE ? 'danger' : 'default';
+  bit.get('userVote') === USER_DOWNVOTE ? 'danger' : 'default';
 
 const getUpvoteButtonStyle = bit =>
-    bit.get('userVote') === USER_UPVOTE ? 'success' : 'default';
+  bit.get('userVote') === USER_UPVOTE ? 'success' : 'default';
