+++ client/src/BitTagPills.jsx
@@ -1,5 +1,5 @@
-import React from "react";
 import { List, Set } from "immutable";
+import React from "react";
 import { Label } from "react-bootstrap";
 
 export default function BitTagPills(props) {
