+++ client/src/Home.jsx
@@ -1,14 +1,14 @@
 import React, { Component } from 'react';
-import { Col, Button } from 'react-bootstrap';
+import { Button, Col } from 'react-bootstrap';
 import { connect } from 'react-redux';
+import * as actionCreators from './action_creators';
 import BitsContainer from './BitsContainer';
 import CreateBitButton from './CreateBitButton';
-import SortingMenu from './SortingMenu';
 import FilterControl from './FilterControl';
 import PageSizeMenu from './PageSizeMenu';
-import * as actionCreators from './action_creators';
-import { getFilters, getSort, getPageSize } from './uri_util';
-import { SORT_DATE, DEFAULT_PAGE_SIZE } from './reducer';
+import { DEFAULT_PAGE_SIZE, SORT_DATE } from './reducer';
+import SortingMenu from './SortingMenu';
+import { getFilters, getPageSize, getSort } from './uri_util';
 
 class Home extends Component {
   constructor(props, context) {
@@ -28,7 +28,7 @@ class Home extends Component {
         <Col md={8}>
           <span>
             <SortingMenu sort={getSort(location.search) || SORT_DATE} />
-            <span style={{float: 'right'}}>
+            <span style={{ float: 'right' }}>
               <PageSizeMenu pageSize={getPageSize(location.search) || DEFAULT_PAGE_SIZE} />
               <Button onClick={() => fetchPreviousPage()}> &lt; </Button>
               <Button onClick={() => fetchNextPage()}> &gt; </Button>
